To avoid circular imports between User and Relationship table, relationship class was added into User model file.
User Model:

Users model functions:
  - Follow other users
  - Get list of their followers
  - Get list of who they are following
  - Unfollow a user
  - Check if useris already following another uesr

Above methods are instance functions on a user object and require username of a second user. Username is argument for second user because it is the id in the user routes, using username avoids requiring to do a database search to run these methods. In this set up, performance can be improved later by refactoring the requests to AJAX requests or by triggering jobs.

Relationship model:
  - Add boolean column 'approved' to differentiate betwen a pending follow request and an approved follower
  - Relationships can be approved
  - A list of unapproved relationships (Follow requests) can be returned